backend/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
djoser==2.2.0
Pillow==9.0.0
django-filter==23.2
drf-extra-fields==3.5.0
psycopg2-binary==2.9.6